Extract KeyIndex from ReflectiveCoreInstance
