+++ legend-pure-runtime/legend-pure-runtime-java-engine-compiled/src/main/java/org/finos/legend/pure/runtime/java/compiled/generation/processors/type/_class/ClassLazyImplProcessor.java
@@ -32,35 +32,34 @@ import org.finos.legend.pure.runtime.java.compiled.generation.processors.type.Ty
 public class ClassLazyImplProcessor
 {
     private static final String IMPORTS = "import java.util.concurrent.atomic.AtomicBoolean;\n" +
+            "import org.eclipse.collections.api.RichIterable;\n" +
+            "import org.eclipse.collections.api.factory.Lists;\n" +
+            "import org.eclipse.collections.api.factory.Maps;\n" +
             "import org.eclipse.collections.api.list.ListIterable;\n" +
             "import org.eclipse.collections.api.list.MutableList;\n" +
-            "import org.eclipse.collections.api.RichIterable;\n" +
-            "import org.eclipse.collections.api.map.MutableMap;\n" +
-            "import org.eclipse.collections.impl.factory.Lists;\n" +
-            "import org.eclipse.collections.impl.factory.Maps;\n" +
-            "import org.eclipse.collections.impl.map.mutable.UnifiedMap;\n" +
             "import org.eclipse.collections.api.map.ImmutableMap;\n" +
+            "import org.eclipse.collections.api.map.MutableMap;\n" +
             "import org.eclipse.collections.impl.list.mutable.FastList;\n" +
-            "import org.finos.legend.pure.m4.coreinstance.CoreInstance;\n" +
-            "import org.finos.legend.pure.m4.coreinstance.factory.CoreInstanceFactory;\n" +
+            "import org.finos.legend.pure.m3.coreinstance.KeyIndex;\n" +
+            "import org.finos.legend.pure.m3.execution.ExecutionSupport;\n" +
             "import org.finos.legend.pure.m4.ModelRepository;\n" +
+            "import org.finos.legend.pure.m4.coreinstance.CoreInstance;\n" +
+            "import org.finos.legend.pure.m4.coreinstance.CoreInstance;\n" +
             "import org.finos.legend.pure.m4.coreinstance.SourceInformation;\n" +
+            "import org.finos.legend.pure.m4.coreinstance.factory.CoreInstanceFactory;\n" +
             "import org.finos.legend.pure.runtime.java.compiled.metadata.MetadataLazy;\n" +
-            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.*;\n" +
-            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.function.defended.*;\n" +
-            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.function.*;\n" +
             "import org.finos.legend.pure.runtime.java.compiled.execution.*;\n" +
             "import org.finos.legend.pure.runtime.java.compiled.execution.sourceInformation.*;\n" +
+            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.*;\n" +
             "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.coreinstance.*;\n" +
-            "import org.finos.legend.pure.runtime.java.compiled.serialization.model.*;\n" +
-            "import org.finos.legend.pure.m3.execution.ExecutionSupport;\n" +
-            "import org.finos.legend.pure.m4.coreinstance.CoreInstance;\n";
+            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.function.*;\n" +
+            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.function.defended.*;\n" +
+            "import org.finos.legend.pure.runtime.java.compiled.serialization.model.*;\n";
 
 
-    private static final String QUALIFIER_IMPORTS =
+    private static final String QUALIFIER_IMPORTS = "import org.eclipse.collections.api.block.function.Function0;\n" +
             "import org.eclipse.collections.api.block.function.Function2;\n" +
-                    "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.map.PureMap;\n" +
-                    "import org.eclipse.collections.api.block.function.Function0;\n";
+            "import org.finos.legend.pure.runtime.java.compiled.generation.processors.support.map.PureMap;\n";
 
     private static final String LAZY_INITIALIZED_SUFFIX = "_initialized";
 
@@ -94,8 +93,8 @@ public class ClassLazyImplProcessor
                 ClassImplProcessor.buildGetValueForMetaPropertyToMany(classGenericType, processorSupport) +
                 ClassImplProcessor.buildSimpleProperties(classGenericType, (property, name, unresolvedReturnType, returnType, returnMultiplicity, returnTypeJava, classOwnerId, ownerClassName, ownerTypeParams, processorContext1) -> "    public final AtomicBoolean _" + name + LAZY_INITIALIZED_SUFFIX + " = new AtomicBoolean(false);\n" +
                         (Multiplicity.isToOne(returnMultiplicity, false) ?
-                                "    public " + returnTypeJava + " _" + name + ";\n" :
-                                "    public RichIterable _" + name + " = Lists.mutable.with();\n") +
+                         "    public " + returnTypeJava + " _" + name + ";\n" :
+                         "    public RichIterable _" + name + " = Lists.mutable.empty();\n") +
                         buildLazyProperty(property, ownerClassName + (ownerTypeParams.isEmpty() ? "" : "<" + ownerTypeParams + ">"), "this", name, returnType, unresolvedReturnType, returnMultiplicity, processorContext1.getSupport(), processorContext1), processorContext, processorSupport) +
                 ClassImplProcessor.buildQualifiedProperties(classGenericType, processorContext, processorSupport) +
                 buildLazyCopy(classGenericType, classInterfaceName, className, false, processorSupport) +
